refactor(nextly): declare the rename target in its own module
The accessor that reads a stored instance's type needs the spelling the
rename moves toward, and inherited the planner's dependencies to get it --
including node:crypto, which is unresolvable in browser code rather than
merely heavy. The declaration now sits apart and the planner re-exports it,
so every existing importer is unaffected and there is one definition.

The accessor's reachable graph is now itself and two constants files.
